ml_application/keras-flask-deploy-webapp-class_pneu/requirements.txt
Flask 
gevent 
h5py 
numpy 
Pillow 
tensorflow-cpu 
Werkzeug
tensorflow_datasets 
